Update reorder_commits splitting and ordering
Makes the sha parsing a little smarter by first separating by the word "pick" in case we capture any other content when we copy the lines.

Removes the reversal of the commits, as it turns out the way we will typically come up with the list of shas (`git rebase -i`) will return the commits in the proper order.